+++ Baseline/Util/FilteredList.cs
@@ -2,12 +2,12 @@ namespace ITI.Baseline.Util;
 
 public class FilteredList<T>
 {
-    public FilteredList(List<T> items, int totalFilteredCount)
+    public FilteredList(T[] items, int totalFilteredCount)
     {
         Items = items;
         TotalFilteredCount = totalFilteredCount;
     }
 
-    public List<T> Items { get; }
+    public T[] Items { get; }
     public int TotalFilteredCount { get; }
 }