+++ DDD/Infrastructure/DataMapping/IMapperExtensions.cs
@@ -1,4 +1,4 @@
-﻿using AutoMapper;
+using AutoMapper;
 using Microsoft.EntityFrameworkCore;
 
 namespace ITI.DDD.Infrastructure.DataMapping;
@@ -10,6 +10,13 @@ public static class IMapperExtensions
         where TDto : class
     {
         return await mapper.ProjectTo<TDto?>(queryable.AsNoTracking()).FirstOrDefaultAsync();
+    }
+
+    public static async Task<TDto[]> ProjectToDtoArrayAsync<T, TDto>(this IMapper mapper, IQueryable<T> queryable)
+        where T : class
+        where TDto : class
+    {
+        return await mapper.ProjectTo<TDto>(queryable.AsNoTracking()).ToArrayAsync();
     }
 
     public static async Task<List<TDto>> ProjectToDtoListAsync<T, TDto>(this IMapper mapper, IQueryable<T> queryable)