+++ DDD/Infrastructure/DataMapping/QueryableExtensions.cs
@@ -1,4 +1,4 @@
-﻿using AutoMapper;
+using AutoMapper;
 
 namespace ITI.DDD.Infrastructure.DataMapping;
 
@@ -9,6 +9,13 @@ public static class QueryableExtensions
         where TDto : class
     {
         return await mapper.ProjectToDtoAsync<TDbEntity, TDto>(q);
+    }
+
+    public static async Task<TDto[]> ProjectToDtoArrayAsync<TDbEntity, TDto>(this IQueryable<TDbEntity> q, IMapper mapper)
+        where TDbEntity : class
+        where TDto : class
+    {
+        return await mapper.ProjectToDtoArrayAsync<TDbEntity, TDto>(q);
     }
 
     public static async Task<List<TDto>> ProjectToDtoListAsync<TDbEntity, TDto>(this IQueryable<TDbEntity> q, IMapper mapper)